# Fix unwanted GPU memory access by CPU