Dereferencing a GPU pointer on the CPU caused significant slowdown in ` MultiBeam::DepositCurrentSlice` (specifically ` a_box_sorter_vec[i].boxOffsetsPtr()[ibox]`). This took about as much time as an entire poisson solve. Now the data is copied to the CPU first with `dtoh_memcpy_async` resulting in a significant speedup (this data specifically is never directly needed on the GPU). This performance issue has been previously unnoticed because the effected code region is only profiled through `Hipace::SolveOneSlice()`.

Runtime in s:

resolution | old | new
-- | -- | --
16 | 0,5214 | 0,3547
32 | 0,8413 | 0,6476
64 | 1,5670 | 1,0980
128 | 2,0120 | 1,7540
256 | 6,8540 | 4,9990
512 | 19,5700 | 13,4200
1024 | 70,1200 | 64,8000


- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [ ] **Tested** (describe the tests in the PR description)
- [ ] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [ ] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [ ] **Constified** (All that can be `const` is `const`)
- [ ] **Code is clean** (no unwanted comments, )
- [ ] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [ ] **Proper label and GitHub project**, if applicable